requirements.txt
aiogram==3.2.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2==2.9.9
Requests==2.31.0
SQLAlchemy==1.4.37
torch[cpu]==2.1.2
transformers==4.36.2